Revert "ファイルを読み込むときに範囲ﾁｪｯｸを追加"
This reverts commit 7d236060b2358b69ab4deb512434011414253f4b.

+++ Tohohu/PlayerShot.h
@@ -11,7 +11,7 @@ struct PlayerShot
 	Vector2 pos;			// À•W
 	Vector2 move;			// ˆÚ“®—Ê
 	float speed;			// ‘¬‚³
-	int moveAngle;			// “®‚­•ûŒü
+	float moveAngle;		// “®‚­•ûŒü
 	bool flag;				// •`‰æÌ×¸Ş
 	float endPos;			// •`‰æI—¹ˆÊ’u
 };
